requirements.txt
faster-whisper==1.0.3
llama-cpp-python==0.2.82
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.24.3
sounddevice==0.4.6
scipy==1.11.4
pyttsx3==2.90
pyaudio==0.2.14
torch==2.1.0
transformers==4.40.0
langchain==0.2.0
langchain-community==0.2.0
chromadb==0.4.24
